requirements.txt
gradio>=3.0
numpy>=1.0
opencv-python>=4.5
pillow>=10.0